Issues-70: Moved RecyclerCustomSwipeActivity to separate module.
