+++ recyclerswipe/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.kirchhoff.recyclerswipe">
+
+    <application>
+        <activity
+            android:name="com.example.recyclerswipe.RecyclerCustomSwipeActivity"
+            android:label="@string/custom_recycler_swipe_title" />
+    </application>
+
+</manifest>
