+++ recyclerswipe/src/main/res/layout/a_recycler_custom_swipe.xml
