+++ recyclerswipe/src/main/res/layout/d_country.xml
