+++ recyclerswipe/src/main/res/layout/r_recycler_custom.xml
